--- nomos-services/data-availability/src/verifier/backend/mod.rs
@@ -1,6 +1,0 @@
-use nomos_core::da::DaVerifier;
-
-pub trait VerifierBackend: DaVerifier {
-    type Settings;
-    fn new(settings: Self::Settings) -> Self;
-}
