--- nomos-services/data-availability/src/verifier/network/adapters/mod.rs
@@ -1,1 +1,0 @@
-
