+++ nomos-services/data-availability/verifier/Cargo.toml
@@ -5,21 +5,11 @@ edition = "2021"
 
 [dependencies]
 async-trait = "0.1"
-blst = "0.3.11"
-bytes = "1.2"
 futures = "0.3"
-kzgrs-backend = { path = "../../../nomos-da/kzgrs-backend" }
 nomos-core = { path = "../../../nomos-core" }
-nomos-da-storage = { path = "../../../nomos-da/storage" }
 nomos-network = { path = "../../../nomos-services/network" }
-nomos-storage = { path = "../../../nomos-services/storage" }
 overwatch-rs = { git = "https://github.com/logos-co/Overwatch", rev = "2f70806" }
 overwatch-derive = { git = "https://github.com/logos-co/Overwatch", rev = "ac28d01" }
-serde = { version = "1.0", features = ["derive"] }
 tokio = { version = "1", features = ["sync", "macros"] }
 tokio-stream = "0.1.15"
 tracing = "0.1"
-
-[features]
-rocksdb-backend = ["nomos-storage/rocksdb-backend"]
-libp2p = ["nomos-network/libp2p"]
