+++ nomos-services/data-availability/verifier/src/backend/mod.rs
@@ -1,5 +1,3 @@
-pub mod kzgrs;
-
 use nomos_core::da::DaVerifier;
 
 pub trait VerifierBackend: DaVerifier {
