+++ nomos-services/data-availability/verifier/src/network/adapters/mod.rs
@@ -1,2 +1,1 @@
-#[cfg(feature = "libp2p")]
-pub mod libp2p;
+
