+++ nomos-services/data-availability/verifier/src/network/mod.rs
@@ -5,7 +5,7 @@ use nomos_network::backends::NetworkBackend;
 use nomos_network::NetworkService;
 use overwatch_rs::services::relay::OutboundRelay;
 use overwatch_rs::services::ServiceData;
-use overwatch_rs::DynError;
+use tokio::sync::oneshot;
 
 #[async_trait::async_trait]
 pub trait NetworkAdapter {
@@ -19,7 +19,7 @@ pub trait NetworkAdapter {
         settings: Self::Settings,
         network_relay: OutboundRelay<<NetworkService<Self::Backend> as ServiceData>::Message>,
     ) -> Self;
-
-    async fn blob_stream(&self) -> Box<dyn Stream<Item = Self::Blob> + Unpin + Send>;
-    async fn send_attestation(&self, attestation: Self::Attestation) -> Result<(), DynError>;
+    async fn blob_stream(
+        &self,
+    ) -> Box<dyn Stream<Item = (Self::Blob, oneshot::Sender<Self::Attestation>)> + Unpin + Send>;
 }
